Reverted back to original database.

+++ EcoPower_Logistics/SuperStore P3/EcoPower_Logistics.Data/Models/OrderDetail.cs
@@ -7,15 +7,15 @@ namespace EcoPower_Logistics.Data.Models
     public partial class OrderDetail
     {
         [DisplayName("Order Details ID")]
-        public short OrderDetailsId { get; set; }
+        public int OrderDetailsId { get; set; }
         [DisplayName("Order ID")]
-        public short OrderId { get; set; }
+        public int OrderId { get; set; }
         [DisplayName("Product ID")]
-        public short ProductId { get; set; }
+        public int ProductId { get; set; }
         [DisplayName("Quantity")]
         public int Quantity { get; set; }
         [DisplayName("Discount")]
-        public int? Discount { get; set; }
+        public double? Discount { get; set; }
 
         public virtual Order Order { get; set; } = null!;
         public virtual Product Product { get; set; } = null!;
